feat(lint): useMediaCaption
