+++ internal/compiler/lint/rules/a11y/useMediaCaption.test.md
@@ -2,13 +2,13 @@
 
 **DO NOT MODIFY**. This file has been autogenerated. Run `rome test internal/compiler/lint/rules/harness.test.ts --update-snapshots` to update.
 
-## `jsx-a11y/useMediaCaption`
+## `a11y/useMediaCaption`
 
 ### `0`
 
 ```
 
- lint/jsx-a11y/useMediaCaption/reject/1/file.tsx:1 lint/jsx-a11y/useMediaCaption ━━━━━━━━━━━━━━━━━━━
+ lint/a11y/useMediaCaption/reject/1/file.tsx:1 lint/a11y/useMediaCaption ━━━━━━━━━━━━━━━━━━━━━━━━━━━
 
   ✖ Provide a track for captions when using audio or video elements.
 
@@ -35,7 +35,7 @@
 
 ```
 
- lint/jsx-a11y/useMediaCaption/reject/2/file.tsx:1 lint/jsx-a11y/useMediaCaption ━━━━━━━━━━━━━━━━━━━
+ lint/a11y/useMediaCaption/reject/2/file.tsx:1 lint/a11y/useMediaCaption ━━━━━━━━━━━━━━━━━━━━━━━━━━━
 
   ✖ Provide a track for captions when using audio or video elements.
 
@@ -62,7 +62,7 @@
 
 ```
 
- lint/jsx-a11y/useMediaCaption/reject/3/file.tsx:1 lint/jsx-a11y/useMediaCaption ━━━━━━━━━━━━━━━━━━━
+ lint/a11y/useMediaCaption/reject/3/file.tsx:1 lint/a11y/useMediaCaption ━━━━━━━━━━━━━━━━━━━━━━━━━━━
 
   ✖ Provide a track for captions when using audio or video elements.
 
@@ -91,7 +91,7 @@
 
 ```
 
- lint/jsx-a11y/useMediaCaption/reject/4/file.tsx:1 lint/jsx-a11y/useMediaCaption ━━━━━━━━━━━━━━━━━━━
+ lint/a11y/useMediaCaption/reject/4/file.tsx:1 lint/a11y/useMediaCaption ━━━━━━━━━━━━━━━━━━━━━━━━━━━
 
   ✖ Provide a track for captions when using audio or video elements.
 
@@ -191,5 +191,110 @@
 <Video>
 	child
 </Video>;
+
+```
+
+### `9`
+
+```
+
+ lint/a11y/useMediaCaption/reject/1/file.html:1 lint/a11y/useMediaCaption ━━━━━━━━━━━━━━━━━━━━━━━━━━
+
+  ✖ Provide a track for captions when using audio or video elements.
+
+    <audio>child</audio>
+    ^^^^^^^^^^^^^^^^^^^
+
+  ℹ Captions support users with hearing-impairments. They should be a transcription or translation
+    of the dialogue, sound effects, musical cues, and other relevant audio information.
+
+━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+
+✖ Found 1 problem
+
+```
+
+### `9: formatted`
+
+```html
+<audio>
+	child
+</audio>
+
+```
+
+### `10`
+
+```
+
+ lint/a11y/useMediaCaption/reject/2/file.html:1 lint/a11y/useMediaCaption ━━━━━━━━━━━━━━━━━━━━━━━━━━
+
+  ✖ Provide a track for captions when using audio or video elements.
+
+    <video>child</video>
+    ^^^^^^^^^^^^^^^^^^^
+
+  ℹ Captions support users with hearing-impairments. They should be a transcription or translation
+    of the dialogue, sound effects, musical cues, and other relevant audio information.
+
+━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+
+✖ Found 1 problem
+
+```
+
+### `10: formatted`
+
+```html
+<video>
+	child
+</video>
+
+```
+
+### `11`
+
+```
+✔ No known problems!
+
+```
+
+### `11: formatted`
+
+```html
+<audio>
+	<track kind="captions" />
+</audio>
+
+```
+
+### `12`
+
+```
+✔ No known problems!
+
+```
+
+### `12: formatted`
+
+```html
+<video>
+	<track kind="captions" />
+</video>
+
+```
+
+### `13`
+
+```
+✔ No known problems!
+
+```
+
+### `13: formatted`
+
+```html
+<video muted>
+</video>
 
 ```
