+++ internal/compiler/lint/rules/a11y/useMediaCaption.test.rjson
@@ -1,0 +1,30 @@
+[
+	{
+		filename: "file.tsx"
+		invalid: [
+			"<audio {...props} />"
+			"<video {...props} />"
+			"<audio>child</audio>"
+			"<video>child</video>"
+	]
+		valid: [
+			"<audio><track kind=\"captions\" {...props} /></audio>"
+			"<video><track kind=\"captions\" {...props} /></video>"
+			"<video muted {...props} ></video>"
+			"<Audio>child</Audio>"
+			"<Video>child</Video>"
+	]
+}
+	{
+		filename: "file.html"
+		invalid: [
+			"<audio>child</audio>"
+			"<video>child</video>"
+	]
+		valid: [
+			"<audio><track kind=\"captions\" /></audio>"
+			"<video><track kind=\"captions\" /></video>"
+			"<video muted ></video>"
+	]
+}
+]
