--- internal/compiler/lint/rules/jsx-a11y/useMediaCaption.test.rjson
@@ -1,14 +1,0 @@
-filename: "file.tsx"
-invalid: [
-	"<audio {...props} />"
-	"<video {...props} />"
-	"<audio>child</audio>"
-	"<video>child</video>"
-]
-valid: [
-	"<audio><track kind='captions' {...props} /></audio>"
-	"<video><track kind='captions' {...props} /></video>"
-	"<video muted {...props} ></video>"
-	"<Audio>child</Audio>"
-	"<Video>child</Video>"
-]
